refactor: refactoring IconProps properties
