+++ src/components/Icon.tsx
@@ -1,27 +1,7 @@
 import React from 'react';
-import { iconSizes } from '../theme/index';
 import { ArrowLeftIcon, CameraIcon, CheckIcon, CompassIcon, DarkIcon, ErrorIcon, HeartIcon, HomeIcon, LightIcon, LockIcon, MobileIcon, NextIcon, PauseIcon, PlayIcon, PrevIcon, SearchIcon, SettingsIcon, ThemeIcon, UserIcon } from '../assets/Icons';
-
-export type IconName = 
-'ArrowLeftIcon' |
-'CameraIcon' |
-'CheckIcon' |
-'CompassIcon' |
-'DarkIcon' |
-'ErrorIcon' |
-'HeartIcon' |
-'HomeIcon' |
-'LightIcon' |
-'LockIcon' |
-'MobileIcon' |
-'NextIcon' |
-'PauseIcon' |
-'PlayIcon' |
-'PrevIcon' |
-'SearchIcon' |
-'SettingsIcon' |
-'ThemeIcon' |
-'UserIcon';
+import { IconProps } from '../assets/Icons/IconProps';
+import { iconSizes, IconName } from '../theme/index';
 
 const Icons = {
   ArrowLeftIcon,
@@ -45,7 +25,7 @@ const Icons = {
   UserIcon
 };
 
-export const Icon = ({name}:{name : IconName}) => {
+export const Icon = ({name, size}:{name : IconName; size: IconProps['size']}) => {
   const Component = Icons[name];
-  return <Component size={iconSizes.sm}/>
+  return <Component size={(iconSizes[size as string])}/>
 }
